// src/app/pages/signup/postulante/postulante-signup-interface.ts
export interface PostulanteSignupRequest {
    nombreUsuario: string;
    apellidoUsuario: string;
    ciudadUsuario: string;
    emailUsuario: string;
    tipodocumentoUsuario: string;
    numerodocumentoUsuario: string;
    usernameUsuario: string;
    passwordUsuario: string;
    generoUsuario: string;
    fechanacimientoUsuario: string

    
}